+++ src/panelWebView/components/Fields/PreviewImageField.tsx
@@ -15,18 +15,21 @@ export interface IPreviewImageFieldProps {
   fieldName: string;
   value: PreviewImageValue | PreviewImageValue[] | null;
   filePath: string | null;
+  parents?: string[];
   multiple?: boolean;
   onChange: (value: string | string[] | null) => void;
 }
 
-export const PreviewImageField: React.FunctionComponent<IPreviewImageFieldProps> = ({label, fieldName, onChange, value, filePath, multiple}: React.PropsWithChildren<IPreviewImageFieldProps>) => {
+export const PreviewImageField: React.FunctionComponent<IPreviewImageFieldProps> = ({label, fieldName, onChange, value, filePath, multiple, parents}: React.PropsWithChildren<IPreviewImageFieldProps>) => {
 
   const selectImage = () => {
     MessageHelper.sendMessage(CommandToCode.selectImage, { 
       filePath: filePath,
       fieldName,
       value,
-      multiple
+      multiple,
+      metadataInsert: true,
+      parents
     });
   };
 
@@ -34,6 +37,8 @@ export const PreviewImageField: React.FunctionComponent<IPreviewImageFieldProps>
     const newValue = value && Array.isArray(value) ? value.filter(image => image.original !== imageToRemove).map(i => i.original) : null;
     onChange(newValue);
   }
+
+  console.log(fieldName, value, filePath, parents)
 
   return (
     <div className={`metadata_field`}>